% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrangePngs.R
\docType{methods}
\name{arrangePngs}
\alias{arrangePngs}
\title{Merge all the rMAUPS results}
\usage{
arrangePngs(outdir)
}
\arguments{
\item{outdir}{Path to rMAUPS results.}
}
\value{
A list of merged figures.
}
\description{
Merge all the rMAUPS results
}
\author{
Wubing Zhang
}
