% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxView.R
\docType{methods}
\name{BoxView}
\alias{BoxView}
\title{Boxplot}
\usage{
BoxView(
  gg,
  x,
  y,
  color = NA,
  fill = NA,
  group = NA,
  width = 0.6,
  size = 1,
  add.jitter = FALSE,
  jitter.color = color,
  jitter.size = size,
  ...
)
}
\arguments{
\item{gg}{A data frame.}

\item{x}{Character, specifying the column name for x plotting.}

\item{y}{Character, specifying the column name for y plotting.}

\item{color}{Character, specifying the column name/color.}

\item{fill}{Character, specifying the column name/fill color.}

\item{group}{Character, specifying the column for grouping.}

\item{width}{Numeric, specifying the box width.}

\item{size}{Numeric, specifying size of the box.}

\item{add.jitter}{Boolean, whether add jitter into the plot.}

\item{jitter.color}{Character, specifying the column name/color of the points.}

\item{jitter.size}{Numeric, specifying size of the jitter.}

\item{...}{Other parameters in geom_boxplot.}
}
\value{
An object created by \code{ggplot}, which can be assigned
and further customized.
}
\description{
Boxplot for a given data frame
}
\author{
Wubing Zhang
}
