% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protSequence.R
\docType{methods}
\name{browseProtStructure}
\alias{browseProtStructure}
\title{View protein sequence positions on protein structure}
\usage{
browseProtStructure(
  protId,
  start,
  end,
  doBrowse = TRUE,
  baseUrl = "https://mupit.icm.jhu.edu/MuPIT_Interactive/?gm=",
 
    checkBaseUrl = "https://mupit.icm.jhu.edu/MuPIT_Interactive/rest/showstructure/check?pos="
)
}
\arguments{
\item{protId}{a uniprot id string}

\item{start}{a vector of start positions}

\item{end}{a vector of end positions}

\item{doBrowse}{a boolean indicator on whether to open browser to view protein structure}

\item{baseUrl}{string for location of mupit service}

\item{checkBaseUrl}{string for location of mupit service to check availability of protein structure}
}
\value{
None
}
\description{
View protein sequence positions on protein structure
}
\author{
Collin Tokheim
}
