% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTM_SEA.R
\name{chopString}
\alias{chopString}
\title{Chops the string to the specified number of characters}
\usage{
chopString(string, nChar = 10, add.dots = T)
}
\arguments{
\item{string}{A number of characters.}

\item{nChar}{N chars.}

\item{add.dots}{Whether add dots into the string.}
}
\description{
Chops the string to the specified number of characters
}
\note{
Given a string and a number of characters
the function chops the string to the
specified number of characters and adds
'...' to the end.
}
