% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSORTdeconv.R
\name{CIBERSORT}
\alias{CIBERSORT}
\title{Runs CIBERSORT for decomposing bulk RNA-seq samples using a single cell reference}
\usage{
CIBERSORT(
  response,
  features = NULL,
  transform = function(x) x,
  usegenes = NULL,
  norm = TRUE,
  nu = c(0.25, 0.5, 0.75),
  optim.nu = FALSE,
  mc.cores = 3,
  ...
)
}
\arguments{
\item{response}{A vector of TPM of bulk RNA-seq samples. Names correspond to genes.}

\item{features}{A TPM matrix representing the gene expression profiles of all cell types that
might contribute to the bulk samples.}

\item{transform}{A function to be applied to columns of exprs following normalization.
Defaults to no transformation since bulk RNA-seq profiles are generated by pooling up
RNA from constituent cell types. In the original CIBERSORT paper, a logarithmic transform was used.}

\item{usegenes}{A vector of genes to be used.}

\item{norm}{Boolean, specifying whether normalize the feature data.}

\item{nu}{Different values of nu to evaluate support vector regression at e1071::svm. Nu defines how
many support vectors (i.e. genes) to use in regression.}

\item{optim.nu}{In the original CIBERSORT implementation, SVR is evaluated at several values of nu and
the value with the best RSME is chosen. This can lead to overfitting. If \code{optim.nu} is set to \code{TRUE},
the value for nu is chosen by cross validation, which leads to longer runtimes.}

\item{mc.cores}{Number of cores used, e.g. for the parallel evaluation at different balues of nu.}

\item{...}{Parameters passed to e1071::svm.}
}
\value{
A data frame in long format suitable for plotting with ggplot2.
}
\description{
This is a custom implementation of the algorithm described by Newman et al (Nautre Methods 12:453-457).
CIBERSORT is an algorithm for estimating the cell type composition of a bulk sample, given a gene expression
profile of the sample and a known gene expression profile for each cell type potentially contributing to the sample.
}
