% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSORTdeconv.R
\name{CIBERSORTdeconv}
\alias{CIBERSORTdeconv}
\title{Runs CIBERSORT for decomposing bulk RNA-seq samples using a single cell reference}
\usage{
CIBERSORTdeconv(exprs, ...)
}
\arguments{
\item{exprs}{A data frame or matrix of raw read counts of bulk RNA-seq samples. Column names correspond to
sample names, row names to genes.}

\item{...}{Parameters passed to e1071::svm.}
}
\value{
A data frame in long format suitable for plotting with ggplot2.
}
\description{
This is a custom implementation of the algorithm described by Newman et al (Nautre Methods 12:453-457).
CIBERSORT is an algorithm for estimating the cell type composition of a bulk sample, given a gene expression
profile of the sample and a known gene expression profile for each cell type potentially contributing to the sample.
}
