% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcFunctions.R
\docType{methods}
\name{corComplex}
\alias{corComplex}
\title{Pairwise correlation of proteins within the same protein complexes}
\usage{
corComplex(dat, idType = c("symbol", "entrez", "uniprot")[1])
}
\arguments{
\item{dat}{A matrix-like object of proteomics data, with gene or protein as row names.}

\item{idType}{A character specifying the protein ID type, one of "symbol" (default), "entrez", "uniprot".}
}
\value{
A data frame including three columns, Protein1, Protein2, and PCC (Pearson Correlation Coefficient).
}
\description{
Pairwise correlation of proteins within the same protein complexes
}
\examples{

}
\author{
Wubing Zhang
}
