% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformCount.R
\docType{methods}
\name{Count2TPM}
\alias{Count2TPM}
\title{Convert read counts to transcripts per million (TPM)}
\usage{
Count2TPM(countMat, idType = "Ensembl", org = "hsa")
}
\arguments{
\item{countMat}{A read count matrix, with geneid as rownames and sample as columns.}

\item{idType}{Type of gene id.}

\item{org}{Organism, hsa or mmu.}
}
\value{
A tpm expression profile.
}
\description{
Convert read counts to transcripts per million (TPM)
}
\author{
Wubing Zhang
}
