% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countNA.R
\docType{methods}
\name{countNA}
\alias{countNA}
\title{Quality control of proteomic data}
\usage{
countNA(expr)
}
\arguments{
\item{expr}{The expression profile.}
}
\value{
a ggplot object.
}
\description{
Quality control of proteomic data
}
\author{
Wubing Zhang
}
