% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcFunctions.R
\docType{methods}
\name{DataCorrelation}
\alias{DataCorrelation}
\title{Correlation between samples/genes in two different data sets.}
\usage{
DataCorrelation(dat1, dat2, method = "spearman")
}
\arguments{
\item{dat1}{A matrix-like object of proteomics data.}

\item{dat2}{A matrix-like object of RNA-Seq data or RPPA data.}

\item{method}{One of "pearson", "kendall", or "spearman" (default).}
}
\value{
A two-column data frame, the first column is the `Correlation`, and the
second column is `Type`, either 'Sample' or 'Gene'.
}
\description{
Correlation between samples/genes in two different data sets.
}
\author{
Wubing Zhang
}
