% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEComplex.R
\docType{methods}
\name{DEComplex}
\alias{DEComplex}
\title{Convert gene level differences to pathway level or complex level differences.}
\usage{
DEComplex(
  depres,
  type = "GOBP+GOCC+CORUM+REACTOME",
  lfc = "log2FC",
  pval = "pvalue"
)
}
\arguments{
\item{depres}{A data frame, with log2FC and pvalue in the column.}

\item{type}{Molecular signatures for testing, available datasets include Pathway
(KEGG, REACTOME, C2_CP), GO (GOBP, GOCC, GOMF), MSIGDB (C1, C2 (C2_CP (C2_CP_PID, C2_CP_BIOCARTA),
C2_CGP), C3 (C3_MIR, C3_TFT), C4, C6, C7, HALLMARK) and Complex (CORUM). Any combination of them
are also accessible (e.g. 'GOBP+GOMF+KEGG+REACTOME').}

\item{lfc}{Gene sets.}

\item{pval}{("PC", default), Pearson, ssGSEA or mean (other value). fisher, stouffer.}
}
\value{
A list.
}
\description{
Convert gene level differences to pathway level or complex level differences.
}
\author{
Wubing Zhang
}
