% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEProtein.R
\docType{methods}
\name{DEProtein}
\alias{DEProtein}
\title{Differential protein abundance analysis with adjustment of differential RNA expression.}
\usage{
DEProtein(
  pData,
  design,
  rData = NULL,
  rData.type = "RNAseq",
  toplabels = NULL,
  ...
)
}
\arguments{
\item{pData}{A matrix of protein abundance profile.}

\item{design}{A vector, indicating the condition of samples in the `pData`.}

\item{rData}{A matrix of gene expression profile.}

\item{rData.type}{RNAseq or array.}

\item{toplabels}{Same as that in the ScatterView.}

\item{...}{Other parameters in ScatterView.}
}
\value{
A list including the differential analysis results and a scatter plot
indicating the adjustment of differential RNA expression from the protein abundance change.
}
\description{
Differential protein abundance analysis with adjustment of differential RNA expression.
}
\author{
Wubing Zhang
}
