% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\docType{methods}
\name{filterN}
\alias{filterN}
\title{Filter out rows with NA or low value}
\usage{
filterN(m, minS = 3, out = NA, impute = "none")
}
\arguments{
\item{m}{A matrix-like object.}

\item{minS}{A numeric, specifying the minimum proportion/number of values should
be quantified for each row.}

\item{out}{A vector, specifying low-quality values (NA and 0).}

\item{impute}{A character, specifying the imputation method, one of none, knn, lls,
lowAbundanceResampling, replicatebasedresampling.}
}
\value{
A matrix with the same columns as input matrix.
}
\description{
Filter out rows with NA or low value
}
\author{
Wubing Zhang
}
