% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcFunctions.R
\name{getDetection}
\alias{getDetection}
\title{Get the number of detected proteins in the given data.}
\usage{
getDetection(dat, samples = NA, condition = NA)
}
\arguments{
\item{dat}{A matrix-like object of proteomics data, in which rows are proteins,
and columns are samples.}

\item{samples}{A character vector, specifying the samples for the analysis.}

\item{condition}{A named character vector, specifying the condition of the samples.}
}
\value{
A data frame, in which the first column includes sample names,
and the second column includes the number of detected proteins.
}
\description{
Get the number of detected proteins in the given data.
}
\author{
Wubing Zhang
}
