% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMERdeconv.R
\name{GetFractions.Abbas}
\alias{GetFractions.Abbas}
\title{Deconvolute the immune cell types.}
\usage{
GetFractions.Abbas(XX, YY, w = NA)
}
\arguments{
\item{XX}{A matrix of immune expression data.}

\item{YY}{A matrix of cancer expression data.}

\item{w}{Weight.}
}
\value{
A list.
}
\description{
Deconvolute the immune cell types.
}
