% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcFunctions.R
\name{getMissing}
\alias{getMissing}
\title{Get the number of missing proteins in the given data.}
\usage{
getMissing(dat, samples = NA, percentage = seq(0, 0.99, 0.01))
}
\arguments{
\item{dat}{A matrix-like object of proteomics data, in which rows are proteins,
and columns are samples.}

\item{samples}{A vector of samples, specifying the samples for the analysis.}
}
\value{
A data frame, in which the first column records the percentage of samples (X),
and the second column includes the number of missing proteins in more than X samples.
}
\description{
Get the number of missing proteins in the given data.
}
\author{
Wubing Zhang
}
