% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProteinAnn.R
\docType{methods}
\name{getProteinAnn}
\alias{getProteinAnn}
\title{Retrieve gene annotations from the NCBI, HNSC, and Uniprot databases.}
\usage{
getProteinAnn(org = "hsa", update = FALSE)
}
\arguments{
\item{org}{Character, hsa (default), bta, cfa, mmu, ptr, rno, ssc are optional.}

\item{update}{Boolean, indicating whether download current annotation.}
}
\value{
A data frame.
}
\description{
Retrieve gene annotations from the NCBI, HNSC, and Uniprot databases.
}
\examples{
ann = getProteinAnn("hsa")
head(ann)

}
\author{
Wubing Zhang
}
