% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTM_SEA.R
\name{gsea.score}
\alias{gsea.score}
\title{Calculate GSEA enrichment score}
\usage{
gsea.score(
  ordered.gene.list,
  gene.set2,
  weight,
  n.rows,
  correl.type,
  gene.set.direction,
  data.expr,
  statistic = c("area.under.RES", "Kolmogorov-Smirnov")[1],
  min.overlap
)
}
\arguments{
\item{ordered.gene.list}{Ordered gene list.}

\item{gene.set2}{Gene sets.}

\item{weight}{Weight as that in ssGSEA2.}

\item{n.rows}{N rows.}

\item{correl.type}{Correlation type as that in ssGSEA2.}

\item{gene.set.direction}{Statistics for testing.}

\item{data.expr}{Expression data matrix.}

\item{statistic}{Statistics for testing.}

\item{min.overlap}{Minimum overlap.}
}
\description{
Calculate GSEA enrichment score
}
