% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\docType{methods}
\name{imputeNA}
\alias{imputeNA}
\title{Imputation}
\usage{
imputeNA(m, method = "knn")
}
\arguments{
\item{m}{A matrix-like object.}

\item{method}{method for imputation, such as knn, lowAbundanceResampling, ReplicateBasedResampling}
}
\value{
A matrix.
}
\description{
Imputation
}
\author{
Wubing Zhang
}
