% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\docType{methods}
\name{lowAbundanceResampling}
\alias{lowAbundanceResampling}
\title{low Abundance Resampling method from protein discover}
\usage{
lowAbundanceResampling(df, percent = 0.05)
}
\arguments{
\item{df}{Matrix-like object.}

\item{percent}{cutoff for low abundance values.}
}
\description{
low Abundance Resampling method from protein discover
}
