% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTM_SEA.R
\name{my.col2rgb}
\alias{my.col2rgb}
\title{Translate a color name into rgb space}
\usage{
my.col2rgb(color, alpha = 80, maxColorValue = 255)
}
\arguments{
\item{color}{A vector of colors.}

\item{alpha}{Alpha.}

\item{maxColorValue}{Max color value.}
}
\description{
Translate a color name into rgb space
}
