% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeProteomics.R
\docType{methods}
\name{normalizeProteomeDiscoverer}
\alias{normalizeProteomeDiscoverer}
\title{Normalization of ProteomeDiscoverer-exported proteomic data}
\usage{
normalizeProteomeDiscoverer(
  prot,
  output = NULL,
  norm = "median",
  log2 = FALSE,
  return = FALSE
)
}
\arguments{
\item{prot}{A file (or folder) path of proteomics data. Or a data frame of proteomics data,
which includes columns of "Contaminant", "Number of unique peptides",
"Gene Symbol", and protein abundance of samples.}

\item{output}{The path to the output file or directory (when \code{prot} is a folder path).}

\item{norm}{method for normalization.}

\item{log2}{Boolean, whether perform log2 transformation.}

\item{return}{Boolean, whether return the data object.}
}
\value{
A data frame.
}
\description{
Normalization of ProteomeDiscoverer-exported proteomic data
}
\author{
Wubing Zhang
}
