% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeProteomics.R
\docType{methods}
\name{normalizeProteomics}
\alias{normalizeProteomics}
\title{Normalize Fisher's data}
\usage{
normalizeProteomics(df, norm = "median", log2 = FALSE)
}
\arguments{
\item{df}{A matrix-like object.}

\item{norm}{method for normalization, such as none, median, medianratio, scale, robustz, quantile, loess.}

\item{log2}{Boolean, whether perform log2 transformation.}
}
\value{
A matrix.
}
\description{
Normalize Fisher's data
}
\author{
Wubing Zhang
}
