% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrattR.R
\name{PrattR}
\alias{PrattR}
\title{Automatically run Pratt and parse the results}
\usage{
PrattR(sequences, api = "/Users/Wubing/Jobs/Project/UPS/_Code/pratt.py",
  minPerc = 20, maxPatternLength = 15,
  maxNumPatternSymbols = maxPatternLength, maxNumWildcard = 9,
  maxNumFlexSpaces = 7, maxFlexibility = 2, maxFlexProduct = 50,
  patternRefinement = TRUE, patternFormat = "PROSITE",
  maxNumPatterns = 100, printVertically = TRUE)
}
\arguments{
\item{sequences}{A vector of sequences.}

\item{api}{The path to pratt.py API.}

\item{minPerc}{the minimum percentage of the input sequences that should match a pattern.}

\item{maxPatternLength}{Maximum pattern length allows you to set the maximum length of a pattern.}

\item{maxNumPatternSymbols}{Maximum number of pattern symbols, PN parameter.}

\item{maxNumWildcard}{Maximum length of a widecard.}

\item{maxNumFlexSpaces}{Maximum number of flexible space.}

\item{maxFlexibility}{The maximum flexibility of a flexible wildcard. For instance x(2,4) and x(10,12) has flexibility 2.}

\item{maxFlexProduct}{Upper limit on the product of a flexibilities for a pattern.}

\item{patternRefinement}{Pattern Refinement.}

\item{patternFormat}{PROSITE Pattern Format. When switched on, patterns
will be output in PROSITE style.}

\item{maxNumPatterns}{Maximum number of patterns between 1 and 100.}

\item{printVertically}{Print vertically. if set, the output is printed vertically instead
of horizontally, vertical output can be better for large sequence sets.}
}
\description{
Automatically run Pratt and parse the results
}
