% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_cptac.R
\name{preprocess_cptac}
\alias{preprocess_cptac}
\title{Standardize the Proteome_CDAP_Protein_Report data}
\usage{
preprocess_cptac(
  CDAP,
  biospecimens.csv = "",
  summary = "",
  outdir = NULL,
  prefix = ""
)
}
\arguments{
\item{CDAP}{A file path, specifying the COAP report data for processing.
You can download the CPTAC data (Proteome_CDAP_Protein_Report or Phosphoproteome_CDAP_Protein_Report)
from https://cptc-xfer.uis.georgetown.edu/publicData/.}

\item{biospecimens.csv}{Path to the biospecimens csv file, with "Aliquot" (corresponding to column names in CDAP)
and "Sample" (Real sample names) in the data.}

\item{summary}{Path to the summary file, which includes counts of the peptide spectrum match.}

\item{outdir}{Output directory.}

\item{prefix}{The prefix for the output files.}
}
\value{
A data frame, representing the stardardized proteome/phosphoproteome data,
in which the first column is median of PSM across multile parallels (if summary is available).
}
\description{
Standardize the Proteome_CDAP_Protein_Report data
}
