% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessCPTAC.R
\docType{methods}
\name{PreprocessCPTAC}
\alias{PreprocessCPTAC}
\title{Preprocess CPTAC data sets}
\usage{
PreprocessCPTAC(exprSet, selection = "Shared", type = "protein")
}
\arguments{
\item{selection}{Unshared or Shared.}

\item{type}{Two choice for data type(protein and phosphosite).
Set type = "protein" to handel the protein abundance data;
Set type = "phosphosite" to handel the phosphorylation site abundance data.}

\item{expr}{Protein abundance data or Phosphorylation site abundance data.}
}
\value{
Uniform protein abundance or phosphorylation site abundance data.
}
\description{
Preprocess CPTAC data sets
}
\author{
Jun Ge
}
