% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMSdata.R
\docType{methods}
\name{processMSdata}
\alias{processMSdata}
\title{Preprocessing MS/MS datasets}
\usage{
processMSdata(proData, rnaData = NULL, id = "Symbol", org = "hsa",
  k = 3)
}
\arguments{
\item{proData}{Proteomics data with rows as gene names and columns as samples}

\item{rnaData}{Relevant RNAseq datasets for better imputation.}

\item{id}{The type of gene ids in the input data.}

\item{org}{Organism.}

\item{k}{An integer.}
}
\value{
A list including QC figures and intermediate results.
}
\description{
Preprocessing MS/MS datasets
}
\author{
Wubing Zhang
}
