% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTM_SEA.R
\name{project.geneset}
\alias{project.geneset}
\title{Executed per gene set}
\usage{
project.geneset(
  data.array,
  gene.names,
  n.cols,
  n.rows,
  weight = 0,
  statistic = "Kolmogorov-Smirnov",
  gene.set,
  nperm = 200,
  correl.type = "rank",
  gene.set.direction = NULL,
  min.overlap,
  size.G.current
)
}
\arguments{
\item{data.array}{Data array}

\item{gene.names}{Gene names.}

\item{n.cols}{N colymns.}

\item{n.rows}{N rows.}

\item{weight}{Weight as that in ssGSEA2.}

\item{statistic}{Statistics for testing.}

\item{gene.set}{Gene sets or signatures.}

\item{nperm}{N permutation.}

\item{correl.type}{Correlation type.}

\item{gene.set.direction}{Gene set direction.}

\item{min.overlap}{Minimum overlap.}

\item{size.G.current}{Size of the current signature.}
}
\description{
Executed per gene set
}
