% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcProteomics.R
\name{qcProteomics}
\alias{qcProteomics}
\title{Quality control of proteomic data}
\usage{
qcProteomics(data, condition = NULL, proj.name = NA, outdir = NULL)
}
\arguments{
\item{data}{A data frame of the proteomics data.}

\item{proj.name}{A character as the prefix of output files.}

\item{outdir}{The path to a local directory.}

\item{meta}{A vector of conditions corresponding to the samples in the proteomics data.}
}
\value{
a list of ggplot objects.
}
\description{
Quality control of proteomic data
}
\author{
Wubing Zhang
}
