% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RabitInput.R
\docType{methods}
\name{RabitInput}
\alias{RabitInput}
\title{From DEanalyze results to Rabit input}
\usage{
RabitInput(deres, idType = "symbol", org = "hsa")
}
\arguments{
\item{deres}{Path to differential expression results or data frame of the DE results.}

\item{idType}{One of "symbol", "ensembl", "uniprot", and "refseq".
the rownames should match colnames in obj, and the first column should be Condition.}

\item{org}{hsa or mmu.}
}
\value{
A weighted gene list.
}
\description{
From DEanalyze results to Rabit input
}
\author{
Wubing Zhang
}
