% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTM_SEA.R
\name{Read.GeneSets.db2}
\alias{Read.GeneSets.db2}
\title{import gene sets into R workspace}
\usage{
Read.GeneSets.db2(gs.db, thres.min = 2, thres.max = 2000)
}
\arguments{
\item{gs.db}{Signature database.}

\item{thres.min}{Minimum size.}

\item{thres.max}{Maxmum size.}
}
\description{
import gene sets into R workspace
}
\note{
20161013 modified by kk
}
