% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protSequence.R
\docType{methods}
\name{readPSPUbiquitin}
\alias{readPSPUbiquitin}
\title{Reads in the ubiquitination site data from Phosphosite Plus database}
\usage{
readPSPUbiquitin(path, minStudies = 1)
}
\arguments{
\item{path}{File path to Ubiquitination_site_dataset from phosphositeplus db}

\item{minStudies}{Minimum number of studies reporting UB site to keep it}
}
\value{
A data frame object
}
\description{
Reads in the ubiquitination site data from Phosphosite Plus database
}
\author{
Collin Tokheim
}
