% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{gsGetter}
\alias{DensityView}
\alias{IdentBarView}
\alias{BarView}
\alias{ScatterView}
\alias{CutoffCalling}
\alias{TransGeneID}
\alias{getGeneAnn}
\alias{getOrtAnn}
\alias{getOrg}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{MAGeCKFlute}{\code{\link[MAGeCKFlute]{BarView}}, \code{\link[MAGeCKFlute]{CutoffCalling}}, \code{\link[MAGeCKFlute]{DensityView}}, \code{\link[MAGeCKFlute]{getGeneAnn}}, \code{\link[MAGeCKFlute]{getOrg}}, \code{\link[MAGeCKFlute]{getOrtAnn}}, \code{\link[MAGeCKFlute]{gsGetter}}, \code{\link[MAGeCKFlute]{IdentBarView}}, \code{\link[MAGeCKFlute]{ScatterView}}, \code{\link[MAGeCKFlute]{TransGeneID}}}
}}

