% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMERdeconv.R
\name{RemoveBatchEffect}
\alias{RemoveBatchEffect}
\title{Combine the cancer expression data and the immune expression data}
\usage{
RemoveBatchEffect(cancer.exp, immune.exp, immune.cellType)
}
\arguments{
\item{cancer.exp}{A data frame or matrix of raw read counts of bulk RNA-seq samples.
Column names correspond to sample names, row names to genes.}

\item{immune.exp}{A matrix of read counts representing the gene expression profiles of
all cell types that might contribute to the bulk samples.}

\item{immune.cellType}{Immune cell types.}
}
\value{
A list.
}
\description{
Combine the cancer expression data and the immune expression data
}
