% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\docType{methods}
\name{replicateBasedResampling}
\alias{replicateBasedResampling}
\title{Replicate based resampling method from protein discover}
\usage{
replicateBasedResampling(df)
}
\arguments{
\item{df}{Matrix-like object.}
}
\description{
Replicate based resampling method from protein discover
}
