% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEqMS.R
\docType{methods}
\name{Residualplot}
\alias{Residualplot}
\title{plot the residuals against the number of quantified peptides/PSMs.}
\usage{
Residualplot(
  fit,
  xlab = "log2(count)",
  ylab = "Variance(fitted - observed)",
  main = NULL
)
}
\arguments{
\item{fit}{an object returned from \code{SpectraCounteBayes} function.}

\item{xlab}{the title for x axis.}

\item{ylab}{the title for y axis.}

\item{main}{the title for the figure.}
}
\value{
a ggplot instance.
}
\description{
plot the residuals against the number of quantified peptides/PSMs.
}
\author{
Yafeng Zhu
}
