% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTM_SEA.R
\name{score}
\alias{score}
\title{Calculate ES score}
\usage{
score(
  max.ES,
  min.ES,
  RES,
  gaps,
  valleys,
  statistic = c("area.under.RES", "Kolmogorov-Smirnov")[1]
)
}
\arguments{
\item{max.ES}{Max enrichment score.}

\item{min.ES}{Min enrichment score.}

\item{RES}{Result.}

\item{gaps}{Gaps}

\item{valleys}{Valleys}

\item{statistic}{Statistics for testing.}
}
\description{
Calculate ES score
}
