% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protSequence.R
\docType{methods}
\name{searchProtSeq}
\alias{searchProtSeq}
\title{Search protein sequences using a regular expression}
\usage{
searchProtSeq(prot_seq_df, regex)
}
\arguments{
\item{prot_seq_df}{A data.frame containing uniprot IDs under the "ID" column and protein sequence under the "protein_sequence" column}

\item{regex}{a string containing the regular expression to search against the protein sequences}
}
\value{
A data frame object
}
\description{
Search protein sequences using a regular expression
}
\author{
Collin Tokheim
}
