% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMERdeconv.R
\name{signatureDetect}
\alias{signatureDetect}
\title{Detection of the cancer type}
\usage{
signatureDetect(exprsn)
}
\arguments{
\item{exprsn}{A data frame or matrix of gene expression profile.}
}
\description{
Detection of the cancer type
}
