% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEqMS.R
\docType{methods}
\name{SpectraCounteBayes}
\alias{SpectraCounteBayes}
\title{Peptide/Spectra Count Based Empirical Bayes Statistics for Differential Expression}
\usage{
SpectraCounteBayes(fit, fit.method = "loess", coef_col)
}
\arguments{
\item{fit}{an list object produced by Limma \code{eBayes} function, it should
have one additional attribute \code{$count}, which stored the peptide or PSM count
quantified for the gene in label-free or isobaric labelled data.}

\item{fit.method}{the method used to fit variance against the number of
peptides/PSM count quantified. Two available methods: "loess","nls" and
"spline". default "loess"."loess" uses \code{loess} and span = 0.75, "nls""
uses a explicit formula \code{y~a+b/x}. "spline" uses \code{smooth.spline} and
"generalized cross-validation" for smoothing parameter computation. For "nls",
independent variable x is peptide/PSM count, response y is pooled variance
(\code{fit$sigma^2}). For "loess" and "spline" method, both x and y are log
transformed before applying the two methods. In most of time, "loess" is sufficient.
To quickly assess the fit model, use \code{VarianceScatterplot} and
\code{Residualplot} functions.}

\item{coef_col}{an integer vector indicating the column(s) of fit$coefficients
for which the function is to be performed. if not specified, all coefficients
are used.}
}
\value{
a list object with the additional attributes being: sca.t - Spectra
Count Adjusted posterior t-value sca.p - Spectra Count Adjusted posterior
p-value sca.dfprior - Spectra Count Adjusted prior degrees of freedom sca.priorvar-
Spectra Count Adjusted estimated prior variance sca.postvar - Spectra Count Adjusted
posterior variance loess.model - fitted loess model.
}
\description{
Peptide/Spectra Count Based Empirical Bayes Statistics for Differential Expression
}
\details{
This function adjusts the T-statistics and p-values for quantitative MS
proteomics experiment according to the number of peptides/PSMs used for
quantification. The method is similar in nature to intensity-based Bayes
method (Maureen A. Sartor et al BMC Bioinformatics 2006).
}
\author{
Yafeng Zhu
}
