% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTM_SEA.R
\name{ssGSEA2}
\alias{ssGSEA2}
\title{Single sample GSEA}
\usage{
ssGSEA2(
  inputMat,
  gene.set.databases = NULL,
  ptmsig.type = c("uniprot", "sitegrpid", "flanking")[1],
  ptmsig.org = c("human", "mouse", "rat")[1],
  sample.norm.type = c("rank", "log", "log.rank", "none")[1],
  weight = 0,
  statistic = c("area.under.RES", "Kolmogorov-Smirnov")[1],
  output.score.type = c("NES", "ES")[1],
  nperm = 1000,
  combine.mode = c("combine.off", "combine.replace", "combine.add")[1],
  min.overlap = 10,
  correl.type = c("rank", "z.score", "symm.rank")[1],
  global.fdr = FALSE,
  extended.output = TRUE,
  par = F,
  spare.cores = 1
)
}
\arguments{
\item{inputMat}{A matrix with rownames matching the terms in ptm signatures.}

\item{gene.set.databases}{List of genesets (in gmt format) to evaluate enrichment on.}

\item{ptmsig.type}{The type of phosphorylation site signature, One of "uniprot", "sitegrpid", or "flanking".}

\item{ptmsig.org}{Organism, "human", "mouse", or "rat".}

\item{sample.norm.type}{Sample normalization, c("rank", "log", "log.rank", "none").}

\item{weight}{when weight==0, all genes have the same weight; if weight>0,
actual values matter, and can change the resulting score.}

\item{statistic}{Method for testing statistic, c("area.under.RES", "Kolmogorov-Smirnov").}

\item{output.score.type}{Output score type, c("NES", "ES")}

\item{nperm}{Number of random permutations for NES case.}

\item{combine.mode}{"combine.off" do not combine *_UP and *_DN versions in a single score.
"combine.replace" combine *_UP and *_DN versions in a single score.
"combine.add" combine *_UP and *_DN versions in a single score and
add it but keeping the individual *_UP and *_DN versions.}

\item{min.overlap}{Minimum overlap genes or phosphorylation sites.}

\item{correl.type}{Correlation type: "rank", "z.score", "symm.rank"}

\item{global.fdr}{If TRUE calculate global FDR; else calculate FDR sample-by-sample}

\item{extended.output}{If TRUE the result GCT files will contain statistics about gene coverage etc.}

\item{par}{Do parallel running.}

\item{spare.cores}{Spare cores.}
}
\description{
Single sample GSEA
}
\note{
Original code written by Pablo Tamayo. Adapted with additional modifications
Referece:
   1. Abazeed, M. E., Adams, D. J., Hurov, K. E., Tamayo, P., Creighton, C. J., Sonkin, D., et al. (2013).
     Integrative Radiogenomic Profiling of Squamous Cell Lung Cancer. Cancer Research, 73(20), 6289–6298.
      http://doi.org/10.1158/0008-5472.CAN-13-1616
   2. Subramanian, A., Tamayo, P., Mootha, V. K., Mukherjee, S., Ebert, B. L., Gillette, M. A., et al. (2005).
      Gene set enrichment analysis: a knowledge-based approach for interpreting genome-wide expression profiles.
     Proceedings of the National Academy of Sciences of the United States of America, 102(43), 15545–15550.
#######################################################################################################
20161013 modified by Karsten Krug
adapt the ssGSEA code to:
  1) work with site specific signature sets
  2) take directionality of regulation into account
  3) multi-threaded using 'doParallel'
  4) handle missing values
  5) cmapR functions for data import/export
  6) produce more extensive output, e.g. dataset/database overlaps, etc.
}
