% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIMERdeconv.R
\name{TIMERdeconv}
\alias{TIMERdeconv}
\title{Runs TIMER for decomposing bulk RNA-seq samples}
\usage{
TIMERdeconv(e, cancer = "AUTO")
}
\arguments{
\item{e}{A data frame or matrix of raw read counts of bulk RNA-seq samples.
Column names correspond to sample names, row names to genes.}

\item{cancer}{A matrix of read counts representing the gene expression profiles of
all cell types that might contribute to the bulk samples.}
}
\description{
Runs TIMER for decomposing bulk RNA-seq samples
}
