% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformCount.R
\docType{methods}
\name{TransformCount}
\alias{TransformCount}
\title{Filter and normalize a count matrix
This wrapper function combines filtering out genes with low reads
in a number of samples with normalization.}
\usage{
TransformCount(
  m,
  method = c("TPM", "voom", "vst", "rlog")[1],
  idType = "Ensembl",
  org = "hsa"
)
}
\arguments{
\item{m}{raw count matrix (genes are rows, columns are samples)}

\item{method}{the normalization method, one of "none", "TPM", "voom", "vst", "rlog".}

\item{idType}{same as in 'Count2TPM', only required when method = "TPM".}

\item{org}{same as in 'Count2TPM', only required when method = "TPM".}
}
\value{
filtered and normalized count matrix
}
\description{
Filter and normalize a count matrix
This wrapper function combines filtering out genes with low reads
in a number of samples with normalization.
}
\author{
Wubing Zhang
}
