% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProteinAnn.R
\docType{methods}
\name{TransProteinID}
\alias{TransProteinID}
\title{Gene ID conversion between protein ids and gene ids}
\usage{
TransProteinID(ids, fromType = "uniprot", toType = "symbol",
  organism = "hsa", ensemblHost = "www.ensembl.org", update = FALSE)
}
\arguments{
\item{ids}{A character vector, input protein identifiers.}

\item{fromType}{The input ID type, one of "uniprot"(default), "refseq", "enst", "symbol".}

\item{toType}{The output ID type, similar to `fromType`.}

\item{organism}{"hsa"(default) or "mmu".}

\item{ensemblHost}{String, specifying ensembl host, you can use `listEnsemblArchives()`
to show all available Ensembl archives hosts.}

\item{update}{Boolean, specifying whether update built-in protein annotation (needs network and takes time).}
}
\value{
A character vector, named by unique input gene ids.
}
\description{
Gene ID conversion between protein ids and gene ids
}
\examples{

}
\seealso{
\code{\link[MAGeCKFlute]{TransGeneID}}
}
\author{
Wubing Zhang
}
