% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEqMS.R
\docType{methods}
\name{VarianceBoxplot}
\alias{VarianceBoxplot}
\title{generate a boxplot of the variance}
\usage{
VarianceBoxplot(
  fit,
  n = 20,
  xlab = "PSM count",
  ylab = "log(Variance)",
  main = NULL
)
}
\arguments{
\item{fit}{an object returned from \code{SpectraCounteBayes} function.}

\item{n}{set a number to plot only the genes with count value smaller or equal to n.}

\item{xlab}{the title for x axis.}

\item{ylab}{the title for y axis.}

\item{main}{the title for the figure.}
}
\value{
a ggplot instance.
}
\description{
generate a boxplot of the variance
}
\author{
Yafeng Zhu
}
