% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEqMS.R
\docType{methods}
\name{VarianceScatterplot}
\alias{VarianceScatterplot}
\title{generate a scatter plot of the variance}
\usage{
VarianceScatterplot(
  fit,
  xlab = "log2(count)",
  ylab = "log(Variance)",
  main = NULL
)
}
\arguments{
\item{fit}{an object returned from \code{SpectraCounteBayes} function.}

\item{xlab}{the title for x axis.}

\item{ylab}{the title for y axis.}

\item{main}{the title for the figure.}
}
\value{
a ggplot instance.
}
\description{
generate a scatter plot of the variance
}
\author{
Yafeng Zhu
}
