% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\docType{methods}
\name{view}
\alias{view}
\title{Visulization of rMAUPS results}
\usage{
view(outdir = "./")
}
\arguments{
\item{outdir}{The path to rMAUPS results.}
}
\value{
Open the shiny app.
}
\description{
Visulization of rMAUPS results
}
\author{
Wubing Zhang
}
