% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corComplex.R
\docType{methods}
\name{corComplex}
\alias{corComplex}
\title{Pairwise correlation of proteins within the same protein complexes}
\usage{
corComplex(data, idType = c("symbol", "entrez", "uniprot")[1])
}
\arguments{
\item{data}{Proteomics data matrix, with gene or protein as row names.}

\item{idType}{The input ID type, one of "symbol" (default), "entrez", "uniprot".}
}
\value{
A ggplot instance.
}
\description{
Pairwise correlation of proteins within the same protein complexes
}
\examples{

}
\author{
Wubing Zhang
}
