% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCorrelation.R
\docType{methods}
\name{DataCorrelation}
\alias{DataCorrelation}
\title{Correlation between RNAseq and Protein}
\usage{
DataCorrelation(ds1, ds2, method = "spearman")
}
\arguments{
\item{ds1}{Path to protein abundance data}

\item{ds2}{Path to RNAseq data}

\item{method}{One of "pearson", "kendall", or "spearman" (default).}
}
\value{
A ggplot instance.
}
\description{
Correlation between RNAseq and Protein
}
\author{
Wubing Zhang
}
