% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deComplex.R
\name{deComplex}
\alias{deComplex}
\title{Convert gene level differences to pathway level or complex level differences.}
\usage{
deComplex(depres, lfc = "log2FC", pval = "pvalue")
}
\arguments{
\item{depres}{normalized count matrix; rows are all genes in the signature that
shall be summarized into one score; columns are samples}

\item{lfc}{Gene sets.}

\item{pval}{("PC", default), Pearson, ssGSEA or mean (other value). fisher, stouffer}
}
\value{
A list.
}
\description{
Convert gene level differences to pathway level or complex level differences.
}
\author{
Wubing Zhang
}
