% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputeNA}
\alias{imputeNA}
\title{Filter out rows with NA or low value}
\usage{
imputeNA(m, method = "knn", k = 30, rowmax = 0.95, colmax = 0.95,
  ...)
}
\arguments{
\item{m}{A matrix-like object.}

\item{method}{method for imputation, such as knn, lowAbundanceResampling, ReplicateBasedResampling}

\item{k}{Integer, parameter for knn.}

\item{rowmax}{parameter for knn.}

\item{colmax}{parameter for knn.}
}
\value{
A matrix.
}
\description{
Filter out rows with NA or low value
}
\author{
Wubing Zhang
}
