% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\docType{methods}
\name{imputeNA}
\alias{imputeNA}
\title{Imputation}
\usage{
imputeNA(m, method = "knn", k = 30, rowmax = 0.95, colmax = 0.95)
}
\arguments{
\item{m}{A matrix-like object.}

\item{method}{method for imputation, such as knn, lowAbundanceResampling, ReplicateBasedResampling}

\item{k}{Integer, parameter for knn.}

\item{rowmax}{parameter for knn.}

\item{colmax}{parameter for knn.}
}
\value{
A matrix.
}
\description{
Imputation
}
\author{
Wubing Zhang
}
