% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAUPSr.R
\docType{methods}
\name{MAUPSr}
\alias{MAUPSr}
\title{rMAUPS pipeline - QC, differential analysis, integrative analysis.}
\usage{
MAUPSr(metadata, outdir = "./", qc = TRUE, type = "msms", method = "limma")
}
\arguments{
\item{metadata}{File path or data frame of the meta data, including columns of
Experiment, Sample, Condition, and multiple comparisons.}

\item{outdir}{Output directory.}

\item{qc}{Boolean, specifying whether perform the quanlity control.}

\item{type}{Could be "msms", "RNASeq", or "Arrary".}

\item{method}{A character, specifying the method for differential analysis.
Optional methods include DESeq2 (RNASeq), GFOLD (RNASeq), limma, glm.pois, glm.qlll, and glm.nb.}
}
\value{
No return value. Output multiple files into local folder.
}
\description{
rMAUPS pipeline - QC, differential analysis, integrative analysis.
}
\author{
Wubing Zhang
}
