% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeProteomics.R
\name{normalizeProteomeDiscoverer}
\alias{normalizeProteomeDiscoverer}
\title{Normalization of ProteomeDiscoverer-exported proteomic data}
\usage{
normalizeProteomeDiscoverer(prot, norm = "median", log2 = FALSE,
  outfile = NULL)
}
\arguments{
\item{prot}{File path (or data frame) of proteomic data,
which includes columns of "Contaminant", "Number of unique peptides",
"Gene Symbol", and protein abundance of samples.}

\item{norm}{method for normalization.}

\item{log2}{Boolean, whether perform log2 transformation.}
}
\value{
A matrix or output a csv file to local folder.
}
\description{
Normalization of ProteomeDiscoverer-exported proteomic data
}
\author{
Wubing Zhang
}
