% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalizeProteomeDiscoverer}
\alias{normalizeProteomeDiscoverer}
\title{Normalize Fisher's data}
\usage{
normalizeProteomeDiscoverer(df.prot, norm = "median", log2 = FALSE)
}
\arguments{
\item{df.prot}{A matrix-like object.}

\item{norm}{method for normalization.}

\item{log2}{Boolean, whether perform log2 transformation.}
}
\value{
A matrix.
}
\description{
Normalize Fisher's data
}
\author{
Wubing Zhang
}
