% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RabitInput.R
\docType{methods}
\name{RabitInput}
\alias{RabitInput}
\title{From DEanalyze results to Rabit input}
\usage{
RabitInput(DEpath, idType = "symbol", org = "hsa", outdir = "./")
}
\arguments{
\item{DEpath}{Path to differential expression results.}

\item{idType}{One of "symbol", "ensembl", "uniprot", and "refseq".
the rownames should match colnames in obj, and the first column should be Condition.}

\item{type}{"Array", "RNASeq" or "msms", only needed when obj is matrix like object.}

\item{method}{Differential expression analysis method, e.g. limma, DESeq2, GFOLD,
glm.pois, glm.qlll, and glm.nb.}

\item{paired}{Boolean, specifying whether perform paired comparison.}

\item{app.dir}{The path to application (e.g. GFOLD).}
}
\value{
An ExpressionSet instance.
}
\description{
From DEanalyze results to Rabit input
}
\seealso{
\code{\link{ExpressionSet-class}}
}
\author{
Wubing Zhang
}
