% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldfunctions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{TransGeneID}
\alias{getGeneAnn}
\alias{EnrichAnalyzer}
\alias{enrich.GSE}
\alias{enrich.ORT}
\alias{enrich.HGT}
\alias{DEAnalyze}
\alias{gsScore}
\alias{BarView}
\alias{ScatterView}
\alias{hclustView}
\alias{cnetplot}
\alias{dotplot}
\alias{emapplot}
\alias{goplot}
\alias{gseaplot}
\alias{heatplot}
\alias{ridgeplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{enrichplot}{\code{\link[enrichplot]{cnetplot}}, \code{\link[enrichplot]{dotplot}}, \code{\link[enrichplot]{emapplot}}, \code{\link[enrichplot]{goplot}}, \code{\link[enrichplot]{gseaplot}}, \code{\link[enrichplot]{heatplot}}, \code{\link[enrichplot]{ridgeplot}}}

  \item{MAGeCKFlute}{\code{\link[MAGeCKFlute]{TransGeneID}}, \code{\link[MAGeCKFlute]{getGeneAnn}}, \code{\link[MAGeCKFlute]{EnrichAnalyzer}}, \code{\link[MAGeCKFlute]{enrich.GSE}}, \code{\link[MAGeCKFlute]{enrich.ORT}}, \code{\link[MAGeCKFlute]{enrich.HGT}}}

  \item{quickAnalyze}{\code{\link[quickAnalyze]{DEAnalyze}}, \code{\link[quickAnalyze]{gsScore}}, \code{\link[quickAnalyze]{BarView}}, \code{\link[quickAnalyze]{ScatterView}}, \code{\link[quickAnalyze]{hclustView}}}
}}

