% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldfunctions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{TransGeneID}
\alias{ScatterView}
\alias{gsGetter}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{MAGeCKFlute}{\code{\link[MAGeCKFlute]{TransGeneID}}, \code{\link[MAGeCKFlute]{ScatterView}}, \code{\link[MAGeCKFlute]{gsGetter}}}
}}

