% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{ReplicateBasedResampling}
\alias{ReplicateBasedResampling}
\title{Replicate based resampling method from protein discover}
\usage{
ReplicateBasedResampling(df)
}
\arguments{
\item{df}{Matrix-like object.}
}
\description{
Replicate based resampling method from protein discover
}
